Fixes variable name reuse in foreach in templates
Some templates used the name of the iterated variable for the item
which does not work anymore.

In most cases replaces `$foo` with `$_foo` for now, as it's also
easily replaceable.

Only in one case actually renames the variable passed to the template
to use a plural which seems like the better option.