Add remote debugging & expose DB port from Docker to host system
Remote debugging will help a lot to understand what's
happening. Xdebug is used which can be used in IntelliJ. A browser
extension will need to be installed, IntelliJ has documentation on how
to do that when creating a PHP Remote Debug run configuration.

Also exposes the DB port to the host system to be able to connect to
it from there.